Refactored home code
